ipd: 64-bit scope ids; the one ARP per SYN at a full backlog stated (D3 review 5, P3)
- ScopeId is a u64: a 32-bit counter panicked (or wrapped) after 2^32
  grants.
- Stated in stack.rs and NAMESPACES: a SYN that finds every listening
  socket busy is refused through the fresh interface, whose empty
  neighbour cache turns the RST into one ARP request for the next hop;
  under a SYN flood at a full backlog that is one ARP broadcast per SYN
  (1:1), never for one of the box's own addresses.

+++ servers/ipd/src/fs.rs
@@ -31,8 +31,9 @@ use crate::link::Netif;
 use crate::scope::Scope;
 use crate::stack::{CtlError, Entropy, Owner, Ready, Stack, WaitFor};
 
-/// A scope's id: never reused.
-pub type ScopeId = u32;
+/// A scope's id: never reused. 64 bits, so no count of grants a box could make wraps it (QA
+/// D3-code-review-5: a 32-bit counter panicked or wrapped after 2^32).
+pub type ScopeId = u64;
 
 /// Where a node is in the tree.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
